MVC refactor
