+++ package.json
@@ -48,7 +48,7 @@
     "nodemon": "^2.0.12",
     "style-loader": "^3.3.1",
     "url-loader": "^4.1.1",
-    "webpack": "^5.50.0",
+    "webpack": "^5.91.0",
     "webpack-cli": "^4.7.2"
   }
 }
