+++ server/routes/usersRoutes.js
@@ -1,0 +1,8 @@
+const express = require ("express");
+const router = express.Router();
+
+const verifyController = require("../controllers/verifyController");
+
+router.post("/verify", verifyController.verify);
+
+module.exports = router;
